+++ src/app/[lang]/page.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { LOCALES } from '@/lib/constants';
 import { getTranslations } from 'next-intl/server';
 import { articleRepository } from '@/infrastructure/article';
-import { ListArticlesUseCase } from '@/application/article/usecase/list-articles.use-case';
+import { ListArticlesUseCase } from '@/application/article/usecase/list-articles.usecase';
 import { ArticleFeaturedFeature } from '@/components/article/browse/ArticleFeaturedFeature';
 import { ArticleCategory } from '@/domain/article/article.metadata';
 import { ArticleSortOption, SortDirection } from '@/domain/article/article.constants';
