+++ src/application/article/usecase/delete-article.usecase.ts
