+++ src/application/article/usecase/get-related-articles.usecase.ts
@@ -3,7 +3,7 @@ import { ArticleSearchResultDto } from '@/application/article/dto/article.dto';
 import { PagedResponse } from '@/domain/shared/pagination';
 import { Article } from '@/domain/article/article';
 import { ArticleStatus } from '@/domain/article/article.control';
-import { SearchArticlesUseCase } from './search-articles.use-case';
+import { SearchArticlesUseCase } from './search-articles.usecase';
 
 import { ArticleCategory } from '@/domain/article/article.metadata';
